shapely/algorithms/min_partition.py
"""
An implementation of the algorithms in:
"Minimum edge length partitioning of rectilinear polygons", by Lingas, Pinter, Rivest and Shamir from 1982
https://people.csail.mit.edu/rivest/pubs/LPRS82.pdf
Programmer: Dvir Borochov
Date: 20-04-24 
"""

def minimum_edge_length_partition(polygon_points):
    """
    The main algorithm.
    Performs minimum edge length partitioning of a rectilinear polygon into rectangles.

    Args:
        polygon_points (list of tuple): List of tuples representing the vertices of the rectilinear polygon.
                                        Each adjacent pair of points has an edge, including between the first and the last point.

    Returns:
        list of list of tuple: A list where each element is a list of tuples representing the vertices of the rectangles
                               resulting from the partition. or None if not found (in case the polygon isn't rectilinear).
                               
    Examples:
        
        >>> minimum_edge_length_partition([(1,5),(1,4),(3,4),(3,3),(2,3),(2,1),(5,1),(8,2),(8,1),(9,1),(9,4),(8,4),(8,5)])
        [[(3,2), (3,3)], [(3,2), (5,2)], [(8,2), (8,4)], [(3,4), (3,5)], [(3,1), (3,2)]]
        
        >>> minimum_edge_length_partition([(2,0),(6,0), (4,6), (4,8), (6,8),(0,6),(0,4), (4,2)])
        [[(2,4), (6,4)]]
        
        >>> minimum_edge_length_partition([(0,0),(0,3),(3,3),(4,3),(4,2),(12,2),(12,10),(10,10),(10,8),(6,8),(6,14),(0,14)])
        [[(0,3), (3,3)], [(4,3), (4,8)], [(0,8), (6,8)], [(10,8),(8,2)], [(10,8),(12,10)]]
    """
    pass

def find_convex_points(polygon_points):
    """
    Finds the convex points in a given rectilinear polygon.

    Args:
        polygon_points (list of tuple): List of tuples representing the vertices of the polygon.

    Returns:
        list: List of tuples representing the convex points of the polygon.

    Examples:
        >>> find_convex_points([(1, 1), (1, 5), (5, 5), (5, 1)])
        [(1, 1), (1, 5), (5, 5), (5, 1)]

        >>> find_convex_points([(2,0),(6,0), (4,6), (4,8), (6,8),(0,6),(0,4), (4,2)])
        [(2,0),(6,0), (4,8), (6,8),(0,6),(0,4)]

        >>> find_convex_points([(0, 0), (0, 3), (1, 3), (1, 1), (3, 1), (3, 0)])
        [(0, 0), (0, 3), (1, 3), (1, 1), (3, 1), (3, 0)]
    """
    pass


def find_matching_point(candidate, polygon):
    """
    Finds a matching point on the induced grid with respect to a given candidate point in a given polygon.

    Args:
        candidate (tuple): The candidate point as a tuple (x, y).
        polygon (list): List of tuples representing the vertices of the polygon.

    Returns:
          list: List of tuples that contains the matching point, or None if not found.
          
     Examples:
        >>> find_matching_point((6, 0), [(2,0),(6,0), (4,6), (4,8), (6,8),(0,6),(0,4), (4,2)])
        [(2,4), (2,6)]

        >>> find_matching_point((8,2), [(1,5),(1,4),(3,4),(3,3),(2,3),(2,1),(5,1),(8,2),(8,1),(9,1),(9,4),(8,4),(8,5)])
        [(3,5), (3,4), (3,3)]
    
    """
    pass


def find_blocked_rectangle(candidate, matching, polygon):
    """
    Finds a blocked rectangle with respect to two points in a given polygon.

    Args:
        candidate (tuple): The first point as a tuple (x, y) which is a candidate point .
        matching (tuple): The second point as a tuple (x, y)which is a matching point in relation to the candidate point
        polygon (list): List of tuples representing the vertices of the polygon.

    Returns:
        list of list of tuple: representing the new edges of the blocked rectangle, or None if not found.
        
    Examples:
        >>> find_blocked_rectangle((6, 0),(2, 4), [(2,0),(6,0), (4,6), (4,8), (6,8),(0,6),(0,4), (4,2)])
        [[(2,4), (6,4)]]

         >>> find_blocked_rectangle((6, 0),(2, 6), [(2,0),(6,0), (4,6), (4,8), (6,8),(0,6),(0,4), (4,2)])
        [[(2,4), (6,4)], [(2,6), (6,6)]]
        
        >>> find_blocked_rectangle((8, 2), (3, 5), [(1,5),(1,4),(3,4),(3,3),(2,3),(2,1),(5,1),(8,2),(8,1),(9,1),(9,4),(8,4),(8,5)])
        [[(3,2), (3,3)], [(3,2), (5,2)], [(8,2), (8,4)], [(3,4), (3,5)]]
    """
    pass


def is_valid_rectilinear_polygon(polygon_points):
    """
    Validates whether the given list of points forms a closed rectilinear polygon.

    Args:
        polygon_points (list of tuple): List of tuples representing the vertices of the polygon.

    Returns:
        bool: True if the points form a valid closed rectilinear polygon, False otherwise.
        
    Examples:
        >>> is_valid_rectilinear_polygon([(2,0),(6,0), (4,6), (4,8), (6,8),(0,6),(0,4), (4,2)])
        True
        
        >>> is_valid_rectilinear_polygon([(1,5),(1,4),(3,4),(3,3),(2,3),(2,1),(5,1),(8,2),(8,1),(9,1),(9,4),(8,4),(8,5)])
        True
              
        >>> is_valid_rectilinear_polygon([(6,0), (4,6), (4,8), (6,8),(0,6),(0,4), (4,2)])
        False    
        
        >>> is_valid_rectilinear_polygon([(1,5),(1,4),(3,4),(3,3),(2,3),(2,1),(5,1),(8,2),(8,1),(9,1),(9,4),(8,4)])
        False   
    """
    return True
