shapely/tests/test_min_partition.py
def test_minimum_edge_length_partition():
    # Test case 1: Rectilinear polygon with minimum edge length partitioning
    polygon_points = [
        (1, 5),
        (1, 4),
        (3, 4),
        (3, 3),
        (2, 3),
        (2, 1),
        (5, 1),
        (8, 2),
        (8, 1),
        (9, 1),
        (9, 4),
        (8, 4),
        (8, 5),
    ]
    assert minimum_edge_length_partition(polygon_points) == [
        [(3, 2), (3, 3)],
        [(3, 2), (5, 2)],
        [(8, 2), (8, 4)],
        [(3, 4), (3, 5)],
        [(3, 1), (3, 2)],
    ]

    # Test case 2: Rectilinear polygon with no minimum edge length partitioning (a square)
    polygon_points = [(0, 0), (4, 0), (4, 4), (0, 4)]
    assert minimum_edge_length_partition(polygon_points) == None

    # Test case 3: Rectilinear polygon with multiple minimum edge length partitionings
    polygon_points = [
        (0, 0),
        (0, 3),
        (3, 3),
        (4, 3),
        (4, 2),
        (12, 2),
        (12, 10),
        (10, 10),
        (10, 8),
        (6, 8),
        (6, 14),
        (0, 14),
    ]
    assert minimum_edge_length_partition(polygon_points) == [
        [(0, 3), (3, 3)],
        [(4, 3), (4, 8)],
        [(0, 8), (6, 8)],
        [(10, 8), (8, 2)],
        [(10, 8), (12, 10)],
    ]

    # Test case 4: Rectilinear polygon with only one minimum edge length partitioning
    polygon_points = [(2, 0), (6, 0), (4, 6), (4, 8), (6, 8), (0, 6), (0, 4), (4, 2)]
    assert minimum_edge_length_partition(polygon_points) == [[(2, 4), (6, 4)]]

    # Test case 5: Rectilinear polygon with no vertices
    polygon_points = []
    assert minimum_edge_length_partition(polygon_points) == None

    # Test case 6: non-Rectilinear polygon
    polygon_points = [(0, 0), (3, 0), (0, 6)]
    assert minimum_edge_length_partition(polygon_points) == None


test_minimum_edge_length_partition()


def test_find_convex_points():
    # Test case 1: Convex polygon
    polygon_points = [(0, 0), (0, 2), (2, 2), (2, 0)]
    assert find_convex_points(polygon_points) == [(0, 0), (0, 2), (2, 2), (2, 0)]

    # Test case 2: Non-convex polygon
    polygon_points = [(0, 0), (0, 2), (1, 1), (2, 2), (2, 0)]
    assert find_convex_points(polygon_points) == [(0, 0), (0, 2), (2, 2), (2, 0)]

    # Test case 3: Convex polygon with duplicate points
    polygon_points = [(0, 0), (0, 2), (2, 2), (2, 0), (0, 0)]
    assert find_convex_points(polygon_points) == [(0, 0), (0, 2), (2, 2), (2, 0)]

    # Test case 4: Convex polygon with collinear points
    polygon_points = [(0, 0), (0, 2), (0, 4), (2, 4), (2, 2), (2, 0)]
    assert find_convex_points(polygon_points) == [(0, 0), (0, 4), (2, 4), (2, 0)]

    # Test case 5: Empty polygon
    polygon_points = []
    assert find_convex_points(polygon_points) == []


test_find_convex_points()


def test_find_matching_point():
    # Test case 1: Matching point exists
    polygon_points = [(2, 0), (6, 0), (4, 6), (4, 8), (6, 8), (0, 6), (0, 4), (4, 2)]
    assert find_matching_point((6, 0), polygon_points) == [(2, 4), (2, 6)]

    # Test case 2: Matching point does not exist
    polygon_points = [
        (1, 5),
        (1, 4),
        (3, 4),
        (3, 3),
        (2, 3),
        (2, 1),
        (5, 1),
        (8, 2),
        (8, 1),
        (9, 1),
        (9, 4),
        (8, 4),
        (8, 5),
    ]
    assert find_matching_point((8, 2), polygon_points) == None

    # Test case 3: Multiple matching points exist
    polygon_points = [
        (1, 5),
        (1, 4),
        (3, 4),
        (3, 3),
        (2, 3),
        (2, 1),
        (5, 1),
        (8, 2),
        (8, 1),
        (9, 1),
        (9, 4),
        (8, 4),
        (8, 5),
    ]
    assert find_matching_point((6, 1), polygon_points) == [(2, 1), (3, 1)]

    # Test case 4: Empty polygon
    polygon_points = []
    assert find_matching_point((2, 0), polygon_points) == None

    # Test case 5: Matching point on the boundary
    polygon_points = [
        (0, 0),
        (0, 3),
        (3, 3),
        (4, 3),
        (4, 2),
        (12, 2),
        (12, 10),
        (10, 10),
        (10, 8),
        (6, 8),
        (6, 14),
        (0, 14),
    ]
    assert find_matching_point((8, 2), polygon_points) == [(3, 5), (3, 4), (3, 3)]


test_find_matching_point()


def test_find_blocked_rectangle():
    # Test case 1: Blocked rectangle exists
    polygon_points = [(2, 0), (6, 0), (4, 6), (4, 8), (6, 8), (0, 6), (0, 4), (4, 2)]
    candidate = (6, 0)
    matching = (2, 4)
    assert find_blocked_rectangle(candidate, matching, polygon_points) == [
        [(2, 4), (6, 4)]
    ]

    # Test case 2: Multiple blocked rectangles exist
    polygon_points = [(2, 0), (6, 0), (4, 6), (4, 8), (6, 8), (0, 6), (0, 4), (4, 2)]
    candidate = (6, 0)
    matching = (2, 6)
    assert find_blocked_rectangle(candidate, matching, polygon_points) == [
        [(2, 4), (6, 4)],
        [(2, 6), (6, 6)],
    ]

    # Test case 3: Blocked rectangle does not exist
    polygon_points = [
        (1, 5),
        (1, 4),
        (3, 4),
        (3, 3),
        (2, 3),
        (2, 1),
        (5, 1),
        (8, 2),
        (8, 1),
        (9, 1),
        (9, 4),
        (8, 4),
        (8, 5),
    ]
    candidate = (8, 2)
    matching = (3, 5)
    assert find_blocked_rectangle(candidate, matching, polygon_points) == None

    # Test case 4: Empty polygon
    polygon_points = []
    candidate = (2, 0)
    matching = (6, 0)
    assert find_blocked_rectangle(candidate, matching, polygon_points) == None

    # Test case 5: Blocked rectangle on the boundary
    polygon_points = [
        (0, 0),
        (0, 3),
        (3, 3),
        (4, 3),
        (4, 2),
        (12, 2),
        (12, 10),
        (10, 10),
        (10, 8),
        (6, 8),
        (6, 14),
        (0, 14),
    ]
    candidate = (8, 2)
    matching = (3, 5)
    assert find_blocked_rectangle(candidate, matching, polygon_points) == [
        [(3, 2), (3, 3)],
        [(3, 2), (5, 2)],
        [(8, 2), (8, 4)],
        [(3, 4), (3, 5)],
    ]


test_find_blocked_rectangle()


def test_is_valid_rectilinear_polygon():
    # Test case 1: Valid rectilinear polygon
    polygon_points = [(2, 0), (6, 0), (4, 6), (4, 8), (6, 8), (0, 6), (0, 4), (4, 2)]
    assert is_valid_rectilinear_polygon(polygon_points) == True

    # Test case 2: Valid rectilinear polygon
    polygon_points = [
        (1, 5),
        (1, 4),
        (3, 4),
        (3, 3),
        (2, 3),
        (2, 1),
        (5, 1),
        (8, 2),
        (8, 1),
        (9, 1),
        (9, 4),
        (8, 4),
        (8, 5),
    ]
    assert is_valid_rectilinear_polygon(polygon_points) == True

    # Test case 3: Invalid rectilinear polygon (not closed)
    polygon_points = [(6, 0), (4, 6), (4, 8), (6, 8), (0, 6), (0, 4), (4, 2)]
    assert is_valid_rectilinear_polygon(polygon_points) == False

    # Test case 4: Invalid rectilinear polygon (not closed)
    polygon_points = [
        (1, 5),
        (1, 4),
        (3, 4),
        (3, 3),
        (2, 3),
        (2, 1),
        (5, 1),
        (8, 2),
        (8, 1),
        (9, 1),
        (9, 4),
        (8, 4),
    ]
    assert is_valid_rectilinear_polygon(polygon_points) == False

    # Test case 5: Invalid rectilinear polygon (not rectilinear)
    polygon_points = [(0, 0), (3, 0), (0, 6)]
    assert is_valid_rectilinear_polygon(polygon_points) == False

    # Test case 6: Invalid rectilinear polygon (not rectilinear)
    polygon_points = [(0, 0), (0, 3), (1, 3), (1, 1), (3, 1), (3, 0)]
    assert is_valid_rectilinear_polygon(polygon_points) == False

    # Test case 7: Empty polygon
    polygon_points = []
    assert is_valid_rectilinear_polygon(polygon_points) == False


test_is_valid_rectilinear_polygon()
